refactor: extract SeriesStats into shared component
- Create new `series-stats` component module in `pages/src/components/`
- Move SeriesStats view and types from discord-series-stats into standalone component
- Refactor discord-series-stats to thin wrapper that delegates to shared component
- Extract StatsHeader component for reusable series/match header pattern
- Update match-stats and series-stats components to use StatsHeader
- Remove discord-series-stats.module.css (styling now in series-stats.module.css)
- Enable SeriesStats component reuse across individual-tracker and other features

The series-stats component now:
- Owns its own CSS and types (series-stats.module.css, types.ts)
- Exports SeriesStatsView for independent use
- Provides re-exported types via discord-series-stats for backwards compatibility
