--- pages/src/components/discord-series-stats/discord-series-stats.module.css
@@ -1,29 +1,0 @@
-.seriesOverviewWrap {
-  overflow-x: auto;
-}
-
-.seriesScoreLink {
-  min-width: 0;
-}
-
-.teamCard {
-  min-width: 140px;
-}
-
-.playerList {
-  width: 100%;
-}
-
-.playerList li {
-  overflow-wrap: anywhere;
-}
-
-.switchView {
-  display: none;
-}
-
-@media (--ultrawide-viewport) {
-  .switchView {
-    display: inline-block;
-  }
-}
