+++ pages/src/components/series-stats/series-stats.module.css
@@ -1,0 +1,206 @@
+.headerBar {
+  display: flex;
+  flex-direction: column;
+  gap: var(--space-3);
+  margin: 0 0 var(--space-6);
+}
+
+.headerLeft {
+  display: flex;
+  flex-direction: column;
+  gap: var(--space-1);
+}
+
+.headerTitle {
+  color: var(--halo-white);
+  font-family: var(--font-heading);
+  font-size: var(--text-2xl);
+  line-height: 1.2;
+  margin: 0;
+}
+
+.headerSubtitle {
+  color: var(--halo-gray);
+  font-family: var(--font-body);
+}
+
+.headerRight {
+  display: flex;
+  flex-direction: column;
+  gap: var(--space-2);
+}
+
+.switchView {
+  display: none;
+}
+
+.dataContainer {
+  background: var(--halo-bg-card);
+  border: 1px solid color-mix(in srgb, var(--halo-teal-primary) 25%, transparent);
+  padding: var(--space-6) 0;
+}
+
+.contentContainer {
+  position: relative;
+  transition: max-width var(--transition-smooth);
+}
+
+.wide {
+  max-width: 100%;
+}
+
+.sectionTitle {
+  color: var(--halo-white);
+  font-family: var(--font-heading);
+  font-size: var(--text-2xl);
+  margin-bottom: var(--space-3);
+  margin-top: 0;
+}
+
+.seriesOverviewWrap {
+  overflow-x: auto;
+}
+
+.seriesOverview {
+  display: flex;
+  flex-direction: column;
+  gap: var(--space-4);
+}
+
+.seriesScores {
+  flex: 1;
+}
+
+.seriesScoresHeader {
+  font-size: 3em;
+  margin: 0;
+  text-align: center;
+}
+
+.seriesScoresList {
+  color: var(--halo-white);
+  display: flex;
+  flex-direction: column;
+  gap: var(--space-4);
+  list-style-type: none;
+  margin: 0;
+  padding: 0;
+}
+
+.seriesScore {
+  background-color: var(--halo-bg-secondary);
+  background-image:
+    linear-gradient(to bottom, var(--series-score-overlay), var(--series-score-overlay)),
+    linear-gradient(
+      to bottom,
+      color-mix(in srgb, var(--team-color, transparent) 20%, transparent),
+      color-mix(in srgb, var(--team-color, transparent) 20%, transparent)
+    ),
+    radial-gradient(circle at 1px 1px, var(--series-score-dot) 1px, transparent 1.5px), var(--series-score-bg);
+  background-position:
+    center,
+    center,
+    0 0,
+    center;
+  background-repeat: no-repeat, no-repeat, repeat, no-repeat;
+  background-size:
+    100% 100%,
+    100% 100%,
+    10px 10px,
+    cover;
+  margin: 0;
+
+  --series-score-bg: none;
+  --series-score-overlay: color-mix(in srgb, var(--halo-bg-secondary) 25%, transparent);
+  --series-score-dot: color-mix(in srgb, var(--halo-white) 15%, transparent);
+}
+
+.seriesScoreLink {
+  align-items: center;
+  border: 2px solid var(--team-color, color-mix(in srgb, var(--halo-teal-primary) 25%, transparent));
+  color: inherit;
+  display: flex;
+  flex-direction: column;
+  font-size: var(--text-3xl);
+  font-weight: 700;
+  height: 100%;
+  justify-content: center;
+  min-width: 0;
+  padding: var(--space-4);
+  position: relative;
+  text-align: center;
+  text-decoration: none;
+  text-shadow:
+    0 1px 2px var(--halo-bg-secondary),
+    0 0 10px var(--halo-bg-secondary);
+}
+
+.seriesSubScore {
+  display: block;
+  font-size: var(--text-xs);
+  font-weight: 400;
+}
+
+.gameTypeIcon {
+  height: 60px;
+  width: 60px;
+}
+
+.gameTypeAndMap {
+  display: block;
+  font-size: var(--text-base);
+}
+
+.teamCard {
+  align-items: center;
+  background: var(--halo-bg-secondary);
+  background-color: color-mix(in srgb, var(--team-color, transparent) 20%, var(--halo-bg-secondary));
+  border: 2px solid var(--team-color, color-mix(in srgb, var(--halo-teal-primary) 25%, transparent));
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  min-width: 140px;
+  padding: var(--space-4);
+  position: relative;
+  text-align: center;
+}
+
+.teamName {
+  color: var(--halo-white);
+  font-family: var(--font-heading);
+  margin: 0 0 var(--space-3);
+}
+
+.playerList {
+  color: var(--halo-white);
+  list-style-type: none;
+  margin: 0;
+  padding: 0;
+  width: 100%;
+}
+
+.playerList li {
+  overflow-wrap: anywhere;
+}
+
+.matchSection {
+  margin-bottom: var(--space-5);
+}
+
+@media (--tablet-viewport) {
+  .headerBar {
+    align-items: flex-start;
+    flex-direction: row;
+    justify-content: space-between;
+  }
+
+  .seriesOverview {
+    flex-direction: row;
+  }
+}
+
+@media (--ultrawide-viewport) {
+  .switchView {
+    display: inline-block;
+  }
+}
